+++ src/ReverseProxy/Service/Management/DestinationManagerFactory.cs
@@ -1,11 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using Microsoft.ReverseProxy.RuntimeModel;
-
 namespace Microsoft.ReverseProxy.Service.Management
 {
     /// <summary>
@@ -15,17 +10,9 @@ namespace Microsoft.ReverseProxy.Service.Management
     /// </summary>
     internal class DestinationManagerFactory : IDestinationManagerFactory
     {
-        private readonly IReadOnlyList<IDestinationChangeListener> _changeListeners;
-
-        public DestinationManagerFactory(IEnumerable<IDestinationChangeListener> changeListeners)
-        {
-            _changeListeners = changeListeners?.ToArray() ?? Array.Empty<IDestinationChangeListener>();
-        }
-
-        /// <inheritdoc/>
         public IDestinationManager CreateDestinationManager()
         {
-            return new DestinationManager(_changeListeners);
+            return new DestinationManager();
         }
     }
 }
