+++ test/ReverseProxy.Tests/Middleware/DestinationInitializerMiddlewareTests.cs
@@ -36,7 +36,7 @@ namespace Microsoft.ReverseProxy.Middleware.Tests
             var httpClient = new HttpMessageInvoker(new Mock<HttpMessageHandler>().Object);
             var cluster1 = new ClusterInfo(
                 clusterId: "cluster1",
-                destinationManager: new DestinationManager(null));
+                destinationManager: new DestinationManager());
             cluster1.ConfigSignal.Value = new ClusterConfig(default, default, default, default, httpClient, default, new Dictionary<string, string>());
             var destination1 = cluster1.DestinationManager.GetOrCreateItem(
                 "destination1",
@@ -78,7 +78,7 @@ namespace Microsoft.ReverseProxy.Middleware.Tests
             var httpClient = new HttpMessageInvoker(new Mock<HttpMessageHandler>().Object);
             var cluster1 = new ClusterInfo(
                 clusterId: "cluster1",
-                destinationManager: new DestinationManager(null));
+                destinationManager: new DestinationManager());
             cluster1.ConfigSignal.Value = new ClusterConfig(
                 new Cluster(),
                 new ClusterHealthCheckOptions(default, new ClusterActiveHealthCheckOptions(enabled: true, Timeout.InfiniteTimeSpan, Timeout.InfiniteTimeSpan, "Any5xxResponse", "")),
